desktop_view/requirements.txt
pillow==11.0.0
qrcode==8.0
pyinstaller==6.11.1
customtkinter==5.2.2